refactor: batch 2
