--- shared/lib/multichain/trustline/index.ts
@@ -1,2 +1,0 @@
-export { isAssetRequireActivate } from './trustline-from-account-asset-info';
-export { getTrustlineAssetPageState } from './get-trustline-asset-page-state';
